Attempt to fix ambiguity with what the 'metadata' of a project is.
Here we have a dedicated variable, called  which
is the basic info we can get from the meta database. getMetadata()
returns the more detailed info as fetched from the API.

Move 'userOptedIn' as a dedicated protected attribute of Project,
and not part of the standard .

Change all instances of 'dbname' to 'dbName' for consistency with other
camel-cased variable names.

Better i18n support for the 'mainspace', using the new new
nsName(ns_id, namespaces) helper to show the namespace name.

Other minor code cleanup