Check for 'www.' prefix on project names
We were already checking for ".org" suffixes, so this just adds
a "www." prefix as well (and checks for them both as well as
singularly).

Also adds the loading spinner to the Project form field and removes
it from the Namespace form field (as the latter loading is triggered
by user action in the former), and adds the spinner to the user and
page select typeaheads too.

Bug: https://phabricator.wikimedia.org/T170763