Base: Use date_format instead of num_format for (c) years in footer (#270)
num_format causes thousands separators to be inserted (so the footer says
"© 2,008-2,019"). 2008 and 2019 are dates, so date_format makes sense and is
semantically correct.

Just using "2008" will cause PHP to interpret the string as an integer, which
gives incorrect output, so "jan-2008" is used instead. This doesn't cause a
visible change since only the year is displayed.