Add IP range support to XTools (excluding Pages Created)
In routing, IP ranges are prefixed with 'ipr-'. This is because Symfony
uses slashes to separate route params, and it'd otherwise could confuse
/sc/192.168.0.1/10 with user 192.168.0.1 in the namespace with ID 10.

Adding IP range support to the Pages tool is much more complex and will
be done in a separate PR. Presumably the Pages tool is not of much
interest for IP ranges, anyway.

Bug: T279041
Bug: T235941