GlobalContribs: show warning when requesting IP and unknown namespace
Because the routing expects namespaces to follow the username, you must
prefix IP ranges with 'ipr-'. This commit shows a warning when the
requested namespace is unknown and the user is a single IP address,
because this most likely means they want to search for the range, which
we link to.