ArticleInfo: catch ServerExceptions for when action API is down
Whenever the Wikimedia action API (or general site outage) happens, the
ServerException from Guzzle is not captured and so users of the
ArticleInfo gadget don't see a response, rather an infinite loading
indicator. In addition, the automated emails with differing titles cause
quite a bit of clutter in the inbox of maintainers.

This commit captures the ServerException, specifically for the
ArticleInfo API endpoint, and fails gracefully.