Indicate when information has been revdel'd or suppressed
This will show '(username removed)' or '(edit summary removed') where
appropriate. For tools that rely on this information, those revisions
are ignored entirely, and we show a flash message warning the user that
the statistics could be inaccurate.

Also fixes a bug in checking whether a revision was marked as minor.