tweak query to handle partially deleted logs (UserRightsRepository.php)
A few things:
 - logs can only be partially revdeleted.
 - log_params is only null when revdeleted.
 - log_params is always null when ''any'' revdeleting was used
 - revdeleted values are null
From that:
A) in userrights.php: we here leave nulls instead of skipping for partially revdeleted logs. This includes added and removed.
B) in userrightsrepository.php: we use LEFT OUTER JOIN and not JOIN actors table, for revdeleted actors. Also remove the IFNULL around comment_text because null is a good signal to twig.
C) in rights_changes_table.html.twig and rights_changes_table.wikitext.twig: if users or comments are null, put a more graceful message.

Bug: T391753