Allow ampersand typing in typeahead for input fields (#519)
Long story short: when the typeahead is open for some reason users can't type &s.

So I added a little listener that blurs and focuses in that case to let the users type the &. Causes a slight half-second blinking of the typeahead, but IMHO it's fine. At least it doesn't just inexplicably prevent typing &s.

Bug: T272851