TopEdits: limit in sql not php to avoid memory issue (#552)
Largely a revert of #547, plus backticks. I then didn't take into account the fact that limiting in PHP not SQL was going to put a boatload of stuff in memory. Was causing memory shortages for TE for users that created many pages. Plus, now we know that the real problem wasn't the assignments (:=) but row_number having become reserved in MariaDB 10.7; therefore, just revert to previous code, and backtick that.

Bug: T396743