PageInfo: remove checks for Wikidata errors
This query has broken multiple times of the years as the Wikidata
database schema is updated for scalability needs. The "errors"
reported were limited to just a missing label and description, and
even then XTools incorrectly did not check the language fallback chain.

If this functionality is to be restored, it should use the API,
per T398932#10984530.

Bug: T398932