TopEdits and PagesCreated: add list of ten top wikiprojects (#521)
DEWISOTT. (Figuring out the queries took a damn long time.)

On performance: same for topedits (just added an instant subquery), and we're at about +1 second for a user with 20K edits for pages. (Had to make new query.) If you've got an idea of how to integrate this into another query, feel free to. Just have bashed my head at SQL for too long today to be in a state to figure it out. Maybe will tomorrow. Didn't do for editcounter, as I don't know where we'd put it; will possibly also do tomorrow.

Bug: T344464