Bump minimum PHP version to 8.2 and update dependencies (#569)
This puts us at the same level of deprecations as we had when running on
PHP 7.4, but we get various other PHP 8 updates. This should us off
until we can manage a full Symfony upgrade.

Best of all, we get to use PHP 8.2 features in our code now! \o/
* https://www.php.net/releases/8.0/en.php
* https://www.php.net/releases/8.1/en.php
* https://www.php.net/releases/8.2/en.php

Bump wikimedia/ip-utils to ^5.0

Bug: T400929