Bump mediawiki/oauthclient to 2.0.* and set user agent in requests
The WMF cluster now requires a user agent, see https://w.wiki/4wJS
This gets set for all Guzzle requests, which is what we should use
instead of file_get_contents() etc. Later when we upgrade Symfony,
we'll move to Symfony's Guzzle-compatible HTTP client.

The host in the Guzzle user agent is hard-coded in YAML
configuration for now.

Bug: T400929