AutoEdits: don't count tools with contribs=true as automated (#563)
Tools with the contribs flag set (eg CXT) are not considered really automated; to be precise, we show them in the non-automated list. This can make the UI confusing, with in extreme cases "0 non-automated edits" but edits in the non-automated list. This PR makes them consistently count as non-automated; thought they still appear in the tools table, renamed from "(Semi-)automated tools" to "Tool counts". (I left the auto-edits message because it's also used in other places.)

Bug: T257723