TopEdits: add summary section, include total edits (#574)
This commit began as just a fix for the show/hide links next to the
namespace headings when viewing TopEdits from an EditCounter result.
Then I noticed TopEdits::getNumPagesAnyNamespace() was being called 22+
times (for each namespace), which is far too inefficient. I think the
more meaningful statistic for the 'all' namespaces view is unique pages
edited across all the namespaces. Conveniently, this query is already
ran as is necessary for pagination, so there's no cost added.

Rename TopEditsRepository::countEditsNamespace() to what it actually is,
::countPagesNamespace(), and give it proper support for 'all'
namespaces. Thus, we no longer need TopEdits::getNumPagesAnyNamespace().

One thing led to another, and we are now showing the familiar "Summary"
section on TopEdits namespace result pages. This also surfaces the
start/end date, which should be there to begin with for clarity and
consistency with other result pages.

Rework TopEdits::getNumTopEdits() to do more what it sounds like: for
TopEdits namespace, it gives the count of total edits. Before this
method was only used for TopEdits per page. The "Total edits" for the
given conditions (namespace, date range) is now exposed. In my tests it
can add a few seconds for extremely prolific users, but usually it's
quite fast.

Finally, fix pagination links to use the start and end dates when
applicable. Also remove the "xt-test--" CSS classes that were apparently
used in integration tests but no longer are.

Bug: T218531