Sign Action API requests on behalf of user when logged in (#567)
Among others, this means that users logging in to XTools will have access to restricted data they (but not everyone) have access to onwiki (eg page watcher count < 30).

Changes (all in Repository/Repository.php):
 - inject requestStack
 - in executeApiRequest, if the user is logged in, use the oauth client and access token in session to make a signed request

Note that passing along a requestStack is optional, so we don't sign in *all* repositories; it's opt-in. Currently it's UserRepo, plus all of those that just directly extended without a custom __construct (which includes PageRepo).

Bug: T401360